Add FRC + computation for age < 18 (#8376)
## Describe your changes
Composer compliance emailed us about this, and we should offer this as a
signal. This is probably better suited for derived DI or configured
somewhere, but in the meantime...

I know we had gone back and forth awhile ago about whether this is a
good one to have because it's based on a theoretically unverified DOB.
However, given we have rules, I think tenants can write their own logic
for this to suit their needs. Therefore, i've just named it with the
"Input" phrasing

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
unit test and tested locally

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->