Add supported document mapping to db_types (#7479)
## Describe your changes
When collecting a document, we need to configure the available
`countries X doc types`. Currently this is managed by hardcoded values
per tenant in `ob_configuration.rs`.

We have several situations we do this in
* must_collect document
* step up (this ones a little sketch)
* redo KYC
* redo Doc

What we would like to do is support:
* requesting specific documents not on the OBC manually
* being able to do "secondary ID" stepups (thanks for uploading your DL,
now upload a passport)

So, ideally:
* We take document and country configurations for documents from the OBC
for normal situations (stepup, must_collect doc) (putting aside stepup
configuration is still a little tbd)
* We support new situations where we "override" the configured OBC value
(like secondary ID stepup, or manually requesting another document) by
saving a document request w/ a mapping of `countries X doc types`. then
when generating the bifrost requirement we can look for an override,
then fall back to OBC.


This PR:
* preps for saving the HashMap<Country, Vec<IdDocKind>> to document
request


## Caveats
It probably makes sense to just move completely away from `OR`ing obc
and document request, but I figured we can punt because 1) the country X
document mapping isn't configured on OBC anyways and 2) it's a little
simpler to figure out where the requirement came from if we `OR`
(suppose we could add another field to doc req) 3) it's easy to migrate
after we do this anyways

## Test Plan


<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->