Modernize signup challenge flow (#8429)
## Describe your changes
During the login flow, we now issue an "identified," unauthed token in `POST /hosted/identify`. This "identified" token is used to make the next requests to initiate the login challenge and to verify the login challenge.

This modernizes the signup challenge flow too - we'll now also issue an "identified," unauthed token  in `POST /hosted/identify/signup_challenge` since we create the vault as soon as the signup flow is initiated.

Now, both the signup flow and login flow can simply pass the auth token to the `POST /hosted/identify/verify` API _instead_ of needing to propagate the ob config auth and sandbox ID all the way through the identify flow. This will let us simplify some frontend and backend code and may in the future let us better handle repeat signup challenges to the same phone/email

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->