Return identified token when there's a conflict in signup challenge (#8431)
## Describe your changes
We're starting to actively prevent making a duplicate vault and force the user to instead log in to their existing account.

We're also updating the API to require an auth token to initiate a login challenge.

So, when there's a conflict (which doesn't happen often, only in race conditions), the client either has to get this auth token from the `POST /hosted/identify` API, or we can just return it inline here in the error. I opted for the latter

## Caveats
This code is a little complex right now to maintain backcompat with clients - will clean this up soon

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->