Fix recent bug in "sign in with email instead" flow (#8448)
## Describe your changes
This bug was introduced in #8430, where we started using the token from
`POST /hosted/identify` to initiate a login challenge. This broke the
KBA flow for "sign in with email instead" because the KBA flow issues a
different token that is needed to initiate the login challenge

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
Tested locally. Will explore some ways to add automated coverage for
this so we don't introduce another regression...