Use country/document field from OBC in the requirements logic (#8438)
## Describe your changes
Now that we have a field on OBC for documents X countries, we can use it if we have it to generate the bifrost requirement.

The issue with alpaca cip is that we currently have logic that looks at residential country to determine document country. This means that users that are in PR _need_ to provide a document from PR.

This is probably a product feature we'd want to guide people to use multiple playbooks in, rather than a single playbook, otherwise configuring various document types for every country is really difficult.

## Caveats
will backfill/clean this code up

## Test Plan
wrote tests

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->