[FP-7072] Institute hard limit number of fp_ids per proxy request
## Describe your changes
Have been monitoring the number of fp_ids per proxy call for the last >1 month. Grid is regularly decrypting 5k at a time, so we'll set this as the max limit

## Caveats
This still can time out when data is cold... so the limit doesn't entirely solve the problems of overly-latent proxy requests. SFTP proxy would help here

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->