Implement partner tenant dashboard auth method (#8454)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->
Guards are coming in a later PR. Just enough here to get a basic
integration test working with the actix extractor. I don't love a lot of
the names in the auth package since there's a lot of similar variants on
`TenantAuth`, but I matched the tenant patterns here and we can think
about a refactor in a later PR.

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->