Use existing identify hooks in confirm KYC data stepup logic (#8527)
## Describe your changes
The confirm page on the collect KYC data machine uses the identify APIs
to allow stepping up an auth token to have permissions to decrypt the
SSN9. It has long had its own copy of hooks for the identify APIs. This
has been a pain in the past - it's hard to remember to update these
little-used hooks when the identify APIs have changes. So, this PR just
migrates the confirm page to use the same hooks as the identify machine

As an aside, i'm a strong believer in the
one-client-method-per-backend-API approach. It's much nicer to update
one hook/typescript request+response and have typescript then show us
which uses need to be updated than to have to grep the codebase for
every corner that may be using an API

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
Tested locally:


https://github.com/onefootprint/monorepo/assets/21066341/aa083706-fbdb-467f-9845-ff5aa9ca410a