Require step-up auth if any required token scopes are missing (#8517)
## Describe your changes
There is an interesting feature called "implicit auth" on the backend.
It is designed to allow users who have signed into a tenant's auth
playbook to not be required to re-auth to sign into the tenant's
onboarding playbook.

The e2e tests for "update login methods" were accidentally using
implicit auth automatically. So the identify machine would see that the
auth token generated for "update login methods" was aready implicitly
logged in and wouldn't require the user to re-authenticate.

But, "update login methods" requires explicit (rather than implicit)
auth in order to update login methods. So, we just add this codified
requirement to the identify machine!

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
Tested locally - token that had scopes but no explicit auth was required
to be stepped up by the identify machine