Support logging in via email for an NYPID (#8526)
## Describe your changes

> When portablizing a NYPID, we would previously only allow logging in
via the unverified phone number, never the unverified email. I believe
this was an arbitrary decision made at the time because most playbooks
require the phone to be verified as a login method. Since that decision
was made, we now have much better handling of a playbok’s required login
methods and will automatically ask the user to verify their phone number
if it is required by a playbook - so I’m thinking we can allow logging
in via email for an NYPID


https://onefootprint.slack.com/archives/C05U1CAD6FQ/p1710192690320129

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
Tested locally.

For an NYPID with both a phone and an email:


https://github.com/onefootprint/monorepo/assets/21066341/27a86f03-2f43-4da4-9911-aa0e28930d79


And an NYPID that only has an email:


https://github.com/onefootprint/monorepo/assets/21066341/855bc721-fa34-4be8-944e-7b94a879665e

In both of these cases, the user is onboarding onto a playbook that
requires phone number as a login method, so they are asked to verify
their phone number after logging in