Create separate API for validation token after identify (#8511)
## Describe your changes
In a new feature, the client will fetch a validation token from the
backend immediately after the user finishes the identify flow. This
allows the tenant to instantly get the fp_id of the user.

We were previously just reusing the `POST /hosted/onboarding/validate`
API - but there are looser constraints for what data exists in the vault
immediately after login. We don't want to relax the constraints of this
normal validate API, so we make one that has looser constraints

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->