Support timeline events for "onboard onto playbook" trigger (#8542)
## Describe your changes
We're adding some more context to an existing timeline event for when
the tenant requests the user to onboard onto a specific playbook. This
is very similar to an old timeline event for when the tenant requests
the user to _reonboard_ onto the playbook they've already onboarded
onto.

For both of these cases, we now add a link to the playbook onto which
the user was requested to onboard.

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
Locally. In this video, I added a sleep to the backend API that fetches
the onboarding config so you can see the loading state better:


https://github.com/onefootprint/monorepo/assets/21066341/a977bfc4-3e6c-46ce-8e7e-c03c61bc9668