Fix bug in obc discovery during login challenge in one-click signup (#8555)
## Describe your changes
For simplicity, the client no longer sends the ob config key in every
identify API request. This is normally fine since the scoped vault
associated with the auth token contains tenant information.

But, when a user is one-click onboarding, we don't create a scoped vault
at the new tenant _until_ the user has logged in. So, there is no scoped
vault associated with the auth token.

We had a bug where we only looked through the scoped vault to find the
tenant ID, but since the scoped vault isn't always populated in
one-click flows, we also need to check the ob config on the auth token

This bug only affected email challenges (which mostly only happen
through the "log in with email instead" flow) because we (arbitrarily)
don't support email challenges without a tenant

https://onefootprint.slack.com/archives/C05SN1UTZLY/p1710345026045809

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
Added integration test. It failed on master w the same error we saw in
prod, and then passed on this branch

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/1ccc5734-0e26-4721-9199-49ecb660b2b4.png)