Add RuleSetVersion table and create method (#8545)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->

Adds a new `rule_set_version` table which will be written to every time
1 or more changes to a playbook's rules is being done (currently rule
additions/deletions/edits are done 1 at a time but we are moving to a
multi-edit flow that will support multiple operations atomically).

Right now Rules are represented in PG in an event-changelog style
pattern:
- 1 user facing rule has a user facing `rule_id`
- There are potentially multiple `rule_instance`'s for a single
`rule_id`, but ever only 1 active (`deactivated_seqno =null`) at a time.
- For updates:
- When a rule is created: we write a new `rule_instance` with a randomly
generated `rule_id`
- When a rule is edited: we set `deactivated_seqno` on the existing
active `rule_instance` and write a new `rule_instance` row with the edit
made.
- When a rule is deleted: we set `deactivated_seqno` on the existing
active `rule_instance` and write a new `rule_instance` row that also has
`deactivated_seqno` set. (kinda odd but sorta nice since it tracks actor
and makes some queries a little simpler)

So the string of unique `rule_instance.created_seqno`'s basically
represent the sequence of seqno's at which different edits to a
playbook's rules were made. This PR adds `rule_set_version` which
provides an explicit/denormalized view of this:
- Whenever any edit(s) are made to rule(s), in addition to the above
`rule_instance` operations, we will also write a `rule_set_version`:
- We follow a similar pattern of always maintaining 1 active
`rule_set_version` per playbook
- When we write a new `rule_set_version` it will have a `version` number
`+1` of that of the current active row.
- We also denormalize `actor` here to represent the actor who made the
edit(s)
- This gives us:
- An easy way to query for all "versions" of a particular playbook's
rules. Pedro is starting to tease designs of this as a user facing
feature and it's something will likely add very soon. (ability to list
all versions, click and view a previous version, revert rules to a
previous version, audit who made what changes when, etc)
- For multi-rule editing, it gives us a solution to concurrent rule
editing for Multi-Rule Editing (main reason this table is being added
right now):
- If two clients are simultaneously making bulk edits to the rules for
the playbooks, then funky write skew problems can occur and given the
huge implications of rules on users/onboardings, it's critical that we
mitigate for this. The solution here is to give the client the current
`rule_set_version.version` for the playbook, when the user is finally
finishing editing various rules and hits `Save changes`, the client
sends up this expected version number along with the intended edits. In
the BE, we can then assert that the version number has not changed since
the user was doing their edits. (if it has we can throw an error, or
possibly get fancier one day and present merge conflicts or auto-merge
or who knows what)

## Caveats
<!--- Are there any caveats to this changeset? -->
The link between `rule_instance`'s and `rule_set_version`'s is still
predicated on seqno's and we are still not strictly modeling the idea of
a "rule set"- a "rule set" currently is just all the rules for a given
playbook. I was thinking it might be nice to start moving in a more
explicit "rule set" direction and actually having a proper join table
between `rule_instance` + `rule_set_version` and have the latter become
more "`rule_set`" (than just _version_).
 
But given current use cases, there isn't an immediate super strong need
for that that I can think of 🤔. The main benefit I can see is if we were
to divorce rules from playbooks and allow mixing and matching different
groups (ie "sets") of rules across playbooks. But until then, I'm
inclined to not introduce new tables/PG rows here? Curious for what
others think tho

## Test Plan
<!--- How did you test your changes? -->
basic unit test

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->