Stop using error JSON wrapper in /hosted/user/challenge (#8581)
## Describe your changes
This error envelope was a remnant of an attempt to return the
`challenge_token` even when an error occurred. But, that still doesn't
really do the trick. Will be more work to support the client receiving
`challenge_token` alongside an error, so will punt till later.

For now, this just removes the error envelope since the client wasn't
organically able to handle it: #8569

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->