Upgrade old instances of aws-actions/configure-aws-credentials to try to fix deployment flake (#8586)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical details and possible side effects. -->
I've been seeing failed prod deployments like this one: https://github.com/onefootprint/monorepo/actions/runs/8286788688/job/22677635868

I'm not exactly sure why the cause is, since the error is a missing image but the image in question shows as pushed in all of our registries before the timestamp of the error. My best guess is that it's an authentication flake or something that's masking a transient error as a 404. I've seen this kind of weirdness from docker credential helpers before... I noticed that the image cache load in question is one of the old instances of `aws-actions/configure-aws-credentials`, so I'm upgrading those to be consistent with the rest.

This doesn't totally make sense though because the logs say the auth worked... worth a try.

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->