Support passkey registration on desktop outside of iframe (#8609)
## Describe your changes
This only starts supporting passkey registration on desktop _for hosted bifrost_, which is run outside of an iframe. This isn't used by many customers today, mostly just for redo flows, so it's not a huge impact. In a future PR, I'll follow up to support passkey registration when bifrost is running inside an iframe

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
Locally:

https://github.com/onefootprint/monorepo/assets/21066341/4d19da20-cd59-401a-a70a-4c5c7be8652c