Refactor `ButtonLink` sizes (#8619)
`ButtonLink` is a component meant to be used with text, even inline in
some cases. Therefore, it'd be easier to allow it to take a text-like
variant instead of a 'size' like buttons. There's no need to predefine a
size --default, compact, small' for this component since the text would
take care of it.

Also removes: 
- sx
- testID 

And uses `Box` as a base to allow all its props