[FP-7347] Unify the state machines for transfer on desktop and mobile (#8617)
## Describe your changes
Right now, the transfer requirement is split into two state machines:
one for mobile and one for desktop.

There are generally three mechanisms for transfering part of a session
to the handoff app, split across mobile/desktop:
1. [mobile] open the handoff session in a new tab. this is only used for
liveness
2. [mobile] send an sms to the phone number on file when the flow has
been opened in a social media browser since some of our flow breaks in
social media browsers
3. [desktop] scan a qr code on (or send an SMS with a link to) your
phone

So, the mobile machine handles (1) + (2), and the desktop machine
handles (3).

In a future PR, we're going to support method (1) in desktop as well. I
started adding (1) to the desktop machine, but this just required
duplicating lots of code. Since there will be good overlap between the
desktop and mobile flows, I decided to just combine these two transfer
machines into one. I basically moved the qr code transfer screen onto
the mobile state machine+router and then consolidated utils/translations

## Caveats
The transfer machine has lots of states - each of the three methods for
transferring has both a main screen and then a "processing" screen. The
"processing" screens can only ever be accessed from the main screens.
Tangential: I think in general in state machine topology, these two
different screens are better represented as sub-screens of a single
state rather than top-level states. In other words, if a state node is
reachable from only one other node, it shouldn't be a top-level state in
the state machine - just a react `useState`. I think this refactor would
make the transfer machine generally easier to understand and reduce the
complexity of the state machine itself

## Test Plan
Conserved unit tests.

Locally, tested each of the 3 methods:

(1) [mobile] open the handoff session in a new tab


https://github.com/onefootprint/monorepo/assets/21066341/19694025-aeba-47b2-b247-36f780d440ce

(2) [mobile] send an sms to the phone number on file when the flow has
been opened in a social media browser


https://github.com/onefootprint/monorepo/assets/21066341/92f86237-17b5-4d4f-be33-3e3e586a3fec

(3) [desktop] scan a qr code on (or send an SMS with a link to) your
phone


https://github.com/onefootprint/monorepo/assets/21066341/b0052e13-32ad-4a3e-8472-c0a275ba708a


and here, we show another subset of this flow: skipping on desktop (when
there's an id doc requirement, which shows an interim confirmation
screen):


https://github.com/onefootprint/monorepo/assets/21066341/6100d571-c9ac-4c34-999b-5c8d44e44592