Add rule_set to OBC wire (#8640)
## Describe your changes

<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->

Adds rule_set to private OBC wire type so FE can pass this up in the
bulk edit API.

## Caveats

<!--- Are there any caveats to this changeset? -->

2 things i went back and forth on:

- Should this be `Option` or not. Theoretically, there is (currently)
always a `rule_set` for every OBC. We create a default set of rules when
we create a playbook. However, we aren't stricly enforcing this in PG
and also things like "Auth" playbooks make me wonder if every kind of
playbook will definitely always have rules. Given that and to just
generally limit the blast radius of a rogue missing `rule_set` erroring
our many other important use cases of OBC, I chose to keep it `Option`
- Should we just join on `rule_set` in `OBC::get` and just return this
everywhere to every caller of this. Lots of callsites don't need this so
its sort of a waste from that perspective but it avoids having to have
the few API call sites that need it to have to do a separate
`RuleSetVersion::get` call I suppose.

Also for the life of me could not figure out the return Type for the
boxed query if I added the join on `rule_set` in `list_query`. Any
Diesel wizardy there very welcome but for now just made a seperate bulk
call to get all rule_set's for a list of obc's

(also still need to rename `RuleSetVersion` -> `RuleSet` sorry for
confusion there)

## Test Plan

<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->