[FP-7334] Support handing off for passkey registration on desktop (#8622)
## Describe your changes
From slack:
https://onefootprint.slack.com/archives/C05KL1TQP9R/p1710473598473649
> On the QR code screen to hand off passkey registration from desktop to
mobile, we show a “Continue on desktop instead” option that just skips
passkey registration altogether. Now that passkeys are generally much
more supported and portable, if the user explicitly selects to continue
on desktop, should we still allow them to register a passkey from their
desktop? This feels better than skipping passkey registration entirely

We weren't able to initially support this when we launched because
passkeys weren't as widely made portable between desktop <> mobile.

## Caveats
We store on the backend, alongside the passkey, the user agent that
registered the passkey, so we can always tell if a passkey was made via
desktop or via mobile. But other than this, it is pretty implicit
whether a passkey was added on desktop or mobile

## Test Plan
Unit tests for the transition from qrTransfer -> newTabRequest

Tested locally:

In an iframe, when we have liveness req, we transfer to new window:

https://github.com/onefootprint/monorepo/assets/21066341/aa2ffda3-ef5d-4048-81a9-74b0fafb5249

Even when there's an id doc requirement, the transfer app on desktop
only handles the liveness requirement:

https://github.com/onefootprint/monorepo/assets/21066341/03094f80-84ea-4520-9274-73a9cda84591

And when we're running liveness in hosted bifrost (not in an iframe), we
handle it inline and don't transfer:


https://github.com/onefootprint/monorepo/assets/21066341/f96d58d8-7ac7-4231-bbe6-353a0d666135