Update Open API documentation (#8647)
## Describe your changes
The docs site renders newlines properly in descriptions of fields. So, this audits some of our descriptions of fields in the API docs and removes newlines

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/ee8f930f-b257-40e9-808c-0acca8dc2831.png)

## Caveats
Perhaps annoying now that there are long lines in code... if this causes us enough pain, we might be able to come up with our own syntax language, like using `\n` in doc strings to mean specifically newline in the docs site. But then, this will cause the IDE docstrings to look ugly

## Test Plan

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/505b6d4a-ea53-4ab6-9b8f-dab4264421f2.png)