Gracefully handle Whatsapp error from invalid recipient when sending OTP (#8678)
## Describe your changes
Before this PR, we would properly fall back to SMS after a failed attempt to send an OTP via whatsapp.

But, there's a case that may be particularly common for which we want a better experience: perhaps a mexican phone number doesn't even have whatsapp account at all - in that case, we shouldn't show any error to the client, we should just silently retry via SMS.

Before this PR, an SMS vendor would return either success or failure. Now, there's a new in between state: "failure that we don't want to show to the client" :)

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
Tested with alex's phone number locally that doesn't have a whatsapp account - had it prefer whatsapp, and then observed fallback behavior. before this PR, we would raise an error message to the client but still retry via SMS, which is confusing. now, we don't show an error message on the client and still fall back to SMS