Use smaller jitter for twilio message status checking retries (#8710)
## Describe your changes
The intent is that we check the status of the message for ~5s to see if it's been sent/delivered.

I noticed when doing some monitoring of whatsapp that sometimes we can only check for ~3s or so, because the jitter could turn a 500ms retry interval into almost a ~0ms retry interval. We need to actually check for 5s for whatsapp because they often take longer to send than whatsapp

So, this changes our jitter strategy to only apply a +- 10% difference, rather than a 0-100% difference

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->